# docs/swagger.yaml
basePath: /api
definitions:
...
  responses.ErrorResponse:
    properties:
      errors: {}
      message:
        type: string
      status:
        type: string
    type: object
  responses.PaginatedResponse:
    properties:
      data: {}
      message:
        type: string
      page:
        type: integer
      page_size:
        type: integer
      status:
        type: string
      total:
        type: integer
      total_pages:
        type: integer
    type: object
  responses.SuccessResponse:
    properties:
      data: {}
      message:
        type: string
      status:
        type: string
    type: object
...
  sport.Equipment:
    properties:
      description:
        type: string
      is_required:
        type: boolean
      name:
        type: string
    type: object
...
  sport.Skill:
    properties:
      created_at:
        type: string
      deleted_at:
        type: string
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      sport_id:
        description: Foreign key to Sport
        type: integer
      updated_at:
        type: string
      weight:
        description: Importance factor of the skill
        type: number
    type: object
  sport.Sport:
    properties:
      created_at:
        type: string
      deleted_at:
        type: string
      description:
        type: string
      equipment:
        items:
          $ref: '#/definitions/sport.Equipment'
        type: array
      icon:
        description: URL or path to icon
        type: string
      id:
        type: integer
      is_active:
        description: To soft-enable/disable sport visibility
        type: boolean
      name:
        type: string
      positions:
        items:
          $ref: '#/definitions/sport.Position'
        type: array
      required_skills:
        description: Skills associated with this sport
        items:
          $ref: '#/definitions/sport.Skill'
        type: array
      rules:
        $ref: '#/definitions/sport.Rules'
      updated_at:
        type: string
    type: object
  sport.UpdateSkillRequest:
    properties:
      description:
        maxLength: 1000
        type: string
      name:
        maxLength: 100
        minLength: 3
        type: string
      weight:
        maximum: 10
        minimum: 0
        type: number
    type: object
...
host: localhost:8088
info:
  contact: {}
  description: "This is a  server for Sport_go\U0001F3CF."
  title: MiowNation REST API(-_-)
  version: "1.0"
paths:
...
  /skills/{skill_id}:
    delete:
      description: Admin can delete a skill
      parameters:
      - description: Skill ID
        in: path
        name: skill_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Skill deleted successfully
          schema:
            $ref: '#/definitions/responses.SuccessResponse'
        "400":
          description: Invalid skill ID
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "404":
          description: Skill not found
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete a skill
      tags:
      - Skills
    put:
      consumes:
      - application/json
      description: Admin can update an existing skill's details
      parameters:
      - description: Skill ID
        in: path
        name: skill_id
        required: true
        type: integer
      - description: Skill update request
        in: body
        name: skill
        required: true
        schema:
          $ref: '#/definitions/sport.UpdateSkillRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/responses.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/sport.Skill'
              type: object
        "400":
          description: Validation error or bad request
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "404":
          description: Skill not found
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "409":
          description: Skill with this name already exists for its sport
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update a skill
      tags:
      - Skills
  /sports:
    get:
      description: Get a list of all available sports with optional filters
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Number of items per page
        in: query
        name: pageSize
        type: integer
      - description: Search term for name or description
        in: query
        name: search
        type: string
      - description: Filter by active status (admin only)
        in: query
        name: is_active
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/responses.PaginatedResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/sport.Sport'
                  type: array
              type: object
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
      summary: Get all sports
      tags:
      - Sports
    post:
      consumes:
      - application/json
      description: Admin can create a new sport
      parameters:
      - description: Sport creation request
        in: body
        name: sport
        required: true
        schema:
          $ref: '#/definitions/sport.CreateSportRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/responses.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/sport.Sport'
              type: object
        "400":
          description: Validation error or bad request
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "409":
          description: Sport with this name already exists
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a new sport
      tags:
      - Sports
  /sports/{sport_id}:
    delete:
      description: Admin can delete a sport (and its associated skills due to DB constraints)
      parameters:
      - description: Sport ID
        in: path
        name: sport_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Sport deleted successfully
          schema:
            $ref: '#/definitions/responses.SuccessResponse'
        "400":
          description: Invalid sport ID
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "404":
          description: Sport not found
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete a sport
      tags:
      - Sports
    get:
      description: Get details of a specific sport by its ID
      parameters:
      - description: Sport ID
        in: path
        name: sport_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/responses.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/sport.Sport'
              type: object
        "404":
          description: Sport not found
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
      summary: Get a sport by ID
      tags:
      - Sports
    put:
      consumes:
      - application/json
      description: Admin can update an existing sport's details
      parameters:
      - description: Sport ID
        in: path
        name: sport_id
        required: true
        type: integer
      - description: Sport update request
        in: body
        name: sport
        required: true
        schema:
          $ref: '#/definitions/sport.UpdateSportRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/responses.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/sport.Sport'
              type: object
        "400":
          description: Validation error or bad request
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "404":
          description: Sport not found
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "409":
          description: Sport with this name already exists (if name changed)
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update a sport
      tags:
      - Sports
...
  /users/me/sports:
    get:
      description: Authenticated user can retrieve their list of sport preferences
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/responses.SuccessResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/sport.UserSport'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get logged-in user's sport preferences
      tags:
      - UserSports
    post:
      consumes:
      - application/json
      description: Authenticated user can add or update their preference for a sport,
        including position and level
      parameters:
      - description: User sport preference request
        in: body
        name: preference
        required: true
        schema:
          $ref: '#/definitions/sport.UserSportRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Preference updated
          schema:
            allOf:
            - $ref: '#/definitions/responses.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/sport.UserSport'
              type: object
        "201":
          description: Preference added
          schema:
            allOf:
            - $ref: '#/definitions/responses.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/sport.UserSport'
              type: object
        "400":
          description: Validation error or bad request
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "404":
          description: Sport not found
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Add or update a sport preference for the logged-in user
      tags:
      - UserSports
...
swagger: "2.0"